Add files via upload
Simply run and it'll ask you for an integer.  The problem with this code is that you HAVE to enter an integer to make this work. I couldn't find a workaround for that. 
PS: If you enter an integer not divisible by 7 or 3, you simply get the message "Have a good day".